Add compatibility with special characters when reading CSVs.
Most of the supported file formats are tied closely to third party software and the time series names are free of special characters. The CSV format is more flexible and the files may come from various sources, contain text in other languages etc.
Therefore it is necessary to add compatibility with special characters such as accents.

Resolves #69