Remove Weibull and Gumbel distribution parameters from returned statistics summary.
These are rarely used in that context and therefore we choose to focus on the end results. Also, it is not straight forward to understand the distribution parameters for the minima distributions. As we flip the sample (*-1), and fit a Weibull distribution to the minima (maxima after *-1), the distribution parameters represent the flipped sample.